# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_web_controllers_http.SwaggerCreateCustomerRequest:
    properties:
      customerCode:
        type: string
      document:
        type: string
      email:
        type: string
      fullName:
        type: string
      notes:
        type: string
      phone:
        type: string
    required:
    - fullName
    type: object
...
  internal_web_controllers_http.SwaggerCustomerListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_domain_entities.Customer'
        type: array
      meta:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerMeta'
    type: object
  internal_web_controllers_http.SwaggerCustomerResponse:
    properties:
      data:
        $ref: '#/definitions/photogallery_api_go_internal_domain_entities.Customer'
    type: object
  internal_web_controllers_http.SwaggerErrorResponse:
    properties:
      code:
        type: string
      detail:
        type: string
      errors:
        items:
          type: string
        type: array
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  internal_web_controllers_http.SwaggerFileListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_domain_entities.File'
        type: array
      meta:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerMeta'
    type: object
  internal_web_controllers_http.SwaggerFileResponse:
    properties:
      data:
        $ref: '#/definitions/photogallery_api_go_internal_domain_entities.File'
    type: object
...
  internal_web_controllers_http.SwaggerMeta:
    properties:
      total:
        type: integer
    type: object
  internal_web_controllers_http.SwaggerOKFlag:
    properties:
      ok:
        type: boolean
    type: object
  internal_web_controllers_http.SwaggerOKFlagResponse:
    properties:
      data:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlag'
    type: object
...
  internal_web_controllers_http.SwaggerUpdateCustomerRequest:
    properties:
      customerCode:
        type: string
      document:
        type: string
      email:
        type: string
      fullName:
        type: string
      notes:
        type: string
      phone:
        type: string
    required:
    - fullName
    type: object
...
  photogallery_api_go_internal_domain_entities.Customer:
    properties:
      createdAtUtc:
        type: string
      customerCode:
        type: string
      document:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      notes:
        type: string
      phone:
        type: string
      updatedAtUtc:
        type: string
    type: object
  photogallery_api_go_internal_domain_entities.File:
    properties:
      contentType:
        type: string
      createdAtUtc:
        type: string
      id:
        type: string
      originalName:
        type: string
      sizeBytes:
        format: int64
        type: integer
      storageBaseKey:
        type: string
      storageKind:
        type: string
      storageRelativePath:
        type: string
      updatedAtUtc:
        type: string
      uploadedAtUtc:
        type: string
      uploadedByUserID:
        type: string
    type: object
...
info:
  contact: {}
  description: API REST v1 Foto-tienda
  title: Photo Gallery API
  version: "1.0"
paths:
...
  /api/v1/customers:
    get:
      description: Lista paginada de clientes con búsqueda y orden opcional.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search text
        in: query
        name: q
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerCustomerListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ListCustomers
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: Crea un nuevo cliente.
      parameters:
      - description: Create customer payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerCreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerCustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: CreateCustomer
      tags:
      - customers
  /api/v1/customers/:id:
    get:
      description: Obtiene un cliente por identificador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerCustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: GetCustomer
      tags:
      - customers
    put:
      consumes:
      - application/json
      description: Actualiza un cliente existente.
      parameters:
      - description: Update customer payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerUpdateCustomerRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerCustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: UpdateCustomer
      tags:
      - customers
  /api/v1/customers/:id/activate:
    patch:
      description: Activa un cliente.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ActivateCustomer
      tags:
      - customers
  /api/v1/customers/:id/deactivate:
    patch:
      description: Desactiva un cliente.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: DeactivateCustomer
      tags:
      - customers
  /api/v1/files:
    get:
      description: Lista paginada de archivos con búsqueda y orden opcional.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search text
        in: query
        name: q
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerFileListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ListFiles
      tags:
      - files
  /api/v1/files/:id:
    delete:
      description: Elimina un archivo por identificador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: DeleteFile
      tags:
      - files
  /api/v1/files/:id/download:
    get:
      description: Descarga un archivo previamente cargado por identificador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: DownloadFile
      tags:
      - files
  /api/v1/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Sube un archivo y lo vincula a una entidad (Sale, WorkOrder, Appointment,
        etc.).
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: Entity type
        in: formData
        name: entityType
        required: true
        type: string
      - description: Entity ID (UUID)
        in: formData
        name: entityId
        required: true
        type: string
      - description: Customer ID (UUID)
        in: formData
        name: customerId
        type: string
      - description: Notes
        in: formData
        name: notes
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: UploadFile
      tags:
      - files
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"